Change namespace of Layer and Bundle controllers
